requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
werkzeug>=3.0.0
google-cloud-vision>=3.4.0

